requirements.txt
Flask==3.1.1
flask-sqlalchemy==3.1.1